[Update] heading text and added ID to article image
The heading text on the index page has been updated from 'Nos Dernières Actus' to 'Nos Derniers Articles'. On the article page, an ID ('articleImage') was added to the main image to enable better manipulation through JavaScript. The 'dataProvider.js' JavaScript file was updated accordingly to handle the image with the new ID.