[Remove] obsolete articles and data provider script
This commit includes the removal of six outdated articles from the repository. The articles were stored in json format, and are no longer applicable to the current context of the project. Consequently, the script used to fetch and fill in the details from these JSON files (dataProvider.js) has also been removed. Additionally, a significant update is made to the header design along with improvements to the layout of the article listings.