[Add] audio feedback to website actions
The code introduces audio feedback for user actions in the website. An audio element has been added to 'index.html' and '404.html' to play specific sounds on different user actions. Additionally, the 'article.js' has been modified to initiate audio playback when an article is clicked. We've also added a new 'click.vtt' file for the audio script.